+++ Sources/BraintreeCore/Public/BraintreeCore/BraintreeCore.h
@@ -8,9 +8,7 @@ FOUNDATION_EXPORT const unsigned char BraintreeCoreVersionString[];
 
 #if __has_include(<Braintree/BraintreeCore.h>)
 #import <Braintree/BTAPIClient.h>
-#import <Braintree/BTAppContextSwitcher.h>
 #import <Braintree/BTBinData.h>
-#import <Braintree/BTClientMetadata.h>
 #import <Braintree/BTClientToken.h>
 #import <Braintree/BTConfiguration.h>
 #import <Braintree/BTEnums.h>
@@ -25,9 +23,7 @@ FOUNDATION_EXPORT const unsigned char BraintreeCoreVersionString[];
 #import <Braintree/BTPreferredPaymentMethodsResult.h>
 #else
 #import <BraintreeCore/BTAPIClient.h>
-#import <BraintreeCore/BTAppContextSwitcher.h>
 #import <BraintreeCore/BTBinData.h>
-#import <BraintreeCore/BTClientMetadata.h>
 #import <BraintreeCore/BTClientToken.h>
 #import <BraintreeCore/BTConfiguration.h>
 #import <BraintreeCore/BTEnums.h>