+++ CHANGELOG.md
@@ -10,6 +10,10 @@
   * BraintreeCore
     * Renamed `BTAppContextSwitchDriver` protocol to `BTAppContextSwitchClient
     * `BTViewControllerPresentingDelegate` protocol now takes in the `client` parameter instead of `driver`
+    * Renamed `BTClientMetadataSourceType` to `BTClientMetadataSource`
+    * Renamed `BTClientMetadataIntegrationType` to `BTClientMetadataIntegration`
+    * Removed static wrapper methods from `BTAppContextSwitcher`
+    * Replaced `BTLogger` with `BTLogLevel` and `BTLogLevelDescription`
   * BraintreeVenmo
     * Renamed `BTVenmoDriver` to `BTVenmoClient`
     * Renamed `BTVenmoDriverErrorDomain` to `BTVenmoErrorDomain`